Added circular enemy pathfinding around the player
Added circular pathfinding system so that when enemies are chasing the player, they target a point around a circle centered on the player instead of going straight for the player. The positions they target around the circle are tied in with how many total enemies there are, and they always choose equidistant points. There is also a WIP shuffling system that will periodically shuffle the assigned positions of the enemies to keep it interesting. I also adjusted many variable values related with attacking, to polish it up a bit.